refactor: update stub
